[FSSDK-10941] event processor files and directories cleanup
