+++ lib/event_processor/batch_event_processor.react_native.spec.ts
@@ -46,13 +46,11 @@ vi.mock('../utils/import.react_native/@react-native-community/netinfo', () => {
 });
 
 import { ReactNativeNetInfoEventProcessor } from './batch_event_processor.react_native';
-import { getMockLogger } from '../tests/mock/mock_logger';
 import { getMockRepeater } from '../tests/mock/mock_repeater';
 import { getMockAsyncCache } from '../tests/mock/mock_cache';
 
 import { EventWithId } from './batch_event_processor';
-import { EventDispatcher } from './event_dispatcher';
-import { formatEvents } from './v1/buildEventV1';
+import { formatEvents } from './event_builder/build_event_v1';
 import { createImpressionEvent } from '../tests/mock/create_event';
 import { ProcessableEvent } from './event_processor';
