+++ lib/event_processor/batch_event_processor.ts
@@ -17,7 +17,7 @@
 import { EventProcessor, ProcessableEvent } from "./event_processor";
 import { Cache } from "../utils/cache/cache";
 import { EventDispatcher, EventDispatcherResponse, EventV1Request } from "./event_dispatcher";
-import { formatEvents } from "../core/event_builder/build_event_v1";
+import { formatEvents } from "./event_builder/build_event_v1";
 import { BackoffController, ExponentialBackoff, IntervalRepeater, Repeater } from "../utils/repeater/repeater";
 import { LoggerFacade } from "../modules/logging";
 import { BaseService, ServiceState, StartupLog } from "../service";
