+++ lib/event_processor/event_builder/build_event_v1.ts
@@ -17,7 +17,7 @@ import {
   EventTags,
   ConversionEvent,
   ImpressionEvent,
-} from '../../event_processor/events';
+} from '../events';
 
 import { Event } from '../../shared_types';
