+++ lib/event_processor/event_builder/event_helpers.tests.js
@@ -18,7 +18,7 @@ import { assert } from 'chai';
 
 import fns from '../../utils/fns';
 import * as projectConfig from '../../project_config/project_config';
-import * as decision from '../decision';
+import * as decision from '../../core/decision';
 import { buildImpressionEvent, buildConversionEvent } from './event_helpers';
 
 describe('lib/event_builder/event_helpers', function() {
